refactor: Move TasksList widget to a separate file and update import paths
